Topic-scoped plans, and the seven defects review found first
A matched topic now becomes a dated plan. A topical day is several
non-contiguous labeled passages, which one range per day cannot say, so
migration 042 adds plan_day_readings (composite PK, cascade, RLS via
EXISTS-on-plans); plan_days keeps the day's lead range so every
range-shaped consumer works unchanged, and book/collection plans simply
have no readings rows. expandTopicalPlan buckets the author's ordered
entries with the same arithmetic chapters use. loadTopic re-verifies the
pointer against the DB — exists, right work, topical_index, published —
so a stale id, a staged work or a forged pointer refuse with a reason.
The builder gains a topic tab: search, pick from suggestions, build.

Then a 3-lens adversarial review ran BEFORE this landed, and it earned
its keep — seven defects, all fixed here, two of them older than this
slice:

G1's digest SQL could not run against EITHER plan table. measureSql
hardcoded `id` while plan_days (shipped hours earlier) and
plan_day_readings key on composite PKs. Red-proved: both raise 42703,
which cutover.mjs reports as "restore from the pre-cutover snapshot" —
a false schema-regression verdict on a healthy database. Neither
classification had ever been executed. idColumns fixes it without
moving any existing digest, and g1-measure-executable.test.ts now runs
the real SQL for every classified table, which nothing did before.

createPlan committed across three transactions, so a transient failure
mid-flight left a topical plan whose days silently understated
themselves. One transaction now.

Also: loadTopic truncated large topics silently while the picker
advertised the true count; a failed topic search claimed the library
was empty; test teardown could strand a published fixture through
empty catches; and readingLabel showed users "Numbers 17:1-999",
leaking the chapter sentinel.

66 plan tests green, residue clean across 20 tables, audit exit 0.

Model: claude-opus-5

+++ web/src/lib/plan/topic-match.ts
@@ -32,7 +32,7 @@ export async function matchTopics(query: string, limit = 3): Promise<TopicMatch[
   // as the INDEXED prefilter; ordering is exact-heading first, then
   // heading-word rank, with body rank only as the tiebreak.
   const rows = (await sql.query(
-    `SELECT s.id AS section_id, s.heading, src.slug AS work_slug, src.title AS work_title,
+    `SELECT s.id::int AS section_id, s.heading, src.slug AS work_slug, src.title AS work_title,
             (SELECT count(*)::int FROM topical_entries te WHERE te.section_id = s.id) AS entry_count
      FROM sections s
      JOIN sources src ON src.id = s.source_id
